pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "justdastit"
version = "3.0.0"
description = "The Burp You Can Afford - Open-source DAST toolkit"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "SnailSploit", email = "kai@snailsploit.com"},
]
keywords = ["security", "pentest", "DAST", "fuzzing", "web-security", "bug-bounty"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Environment :: Console",
    "Framework :: AsyncIO",
]

dependencies = [
    "httpx>=0.25.0",
    "click>=8.1.0",
    "rich>=13.0",
]

[project.optional-dependencies]
proxy = ["mitmproxy>=10.0"]
reports = ["jinja2>=3.1"]
mcp = ["fastmcp>=2.0"]
config = ["pyyaml>=6.0"]
shell = ["prompt-toolkit>=3.0"]
browser = ["playwright>=1.40"]
full = [
    "mitmproxy>=10.0",
    "jinja2>=3.1",
    "fastmcp>=2.0",
    "pyyaml>=6.0",
    "prompt-toolkit>=3.0",
    "playwright>=1.40",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "mypy>=1.5",
    "ruff>=0.1",
    "respx>=0.20",
]

[project.scripts]
justdastit = "justdastit.cli:main"

[project.urls]
Homepage = "https://github.com/snailsploit/justdastit"
Documentation = "https://snailsploit.com/justdastit"
Repository = "https://github.com/snailsploit/justdastit"
Issues = "https://github.com/snailsploit/justdastit/issues"

[tool.setuptools.packages.find]
include = ["justdastit*"]

[tool.setuptools.package-data]
justdastit = ["py.typed", "wordlists/*.txt", "templates/*.html"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
